Input form validation of product and order
- create form input validation for product and order (use custom validators for number, price, imageurl - see validators.ts)
- error handling
- put RBAC for order in sidebar menu

+++ src/app/pages/product/product.component.ts
@@ -1,9 +1,10 @@
-import { Component, ViewChild, OnInit, ChangeDetectorRef } from '@angular/core';
+import { Component, ViewChild, OnInit, ChangeDetectorRef, ElementRef } from '@angular/core';
 import { ModalConfig, ModalComponent } from '../../_metronic/partials';
 import { SharedService } from 'src/app/shared.service';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Observable } from 'rxjs';
 import { AuthService } from 'src/app/modules/auth';
+import { ImageUrlValidator, NumberValidator, PriceValidator } from '../validators';
 
 @Component({
   selector: 'app-product',
@@ -17,6 +18,7 @@ export class ProductComponent implements OnInit {
     closeButtonLabel: 'Cancel'
   };
   @ViewChild('modal') private modalComponent: ModalComponent;
+  @ViewChild('productFormElem') productFormElem: ElementRef  | undefined;
   constructor(private service:SharedService, private fb: FormBuilder, private authService:AuthService, private cdr: ChangeDetectorRef) {}
 
   products$: Observable<any[]>;
@@ -31,15 +33,17 @@ export class ProductComponent implements OnInit {
   }
 
   ngOnInit() {
+
     this.refreshProducts();
     this.loadCategories();
+
     this.productForm = this.fb.group({
-      category: [''],
-      description: [''],
-      imageUrl: [''],
-      name: [''],
-      price: [''],
-      stockQuantity: ['']
+      category: ['', Validators.required],
+      description: ['', [Validators.required, Validators.maxLength(100)]],
+      imageUrl: ['', [Validators.required, ImageUrlValidator()] ], // imageUrl: only accepts valid url with extensions 'jpg', 'jpeg', 'png', 'gif'
+      name: ['', [Validators.required, Validators.maxLength(30)]],
+      price: ['', [Validators.required, PriceValidator()]], // price: regex - ^\d+(\.\d{1,2})?$/ (number with up to 2 decimal points)
+      stockQuantity: ['', [Validators.required, NumberValidator()]] // stockQuantity: only accept positive number
     });
   }
 
@@ -61,10 +65,24 @@ export class ProductComponent implements OnInit {
   }
 
   deleteProduct(id:string) {
-    this.service.deleteProduct(id).then((res)=>{
-      console.log(res);
-      this.refreshProducts();
-    })
+    this.errorMessage = null;
+  
+    this.authService.canPerformAction('admin').then(canPerform => {
+      if (!canPerform) {
+        this.displayErrorAlert('Unauthorized: Insufficient permissions to perform this action.');
+        return;
+      } else {
+        this.service.deleteProduct(id).then((res)=>{
+          this.refreshProducts();
+        }).catch(error => {
+          console.error('Error deleting product:', error);
+          this.displayErrorAlert('Error in deleting product. Please try again.');
+        });
+      }
+    }).catch(error => {
+      console.error('Error checking permissions:', error);
+      this.displayErrorAlert('Error checking permissions. Please refresh the page or try again later.');
+    });
   }
 
   filterProducts(event: Event) {
@@ -94,6 +112,8 @@ export class ProductComponent implements OnInit {
   
     // Assuming you have a property to hold the ID of the product being edited
     this.editingProductId = product.id;
+
+    this.focusOnForm();
   }
   
   saveProduct() {
@@ -102,35 +122,37 @@ export class ProductComponent implements OnInit {
   
     this.authService.canPerformAction('admin').then(canPerform => {
       if (!canPerform) {
-        console.error('Unauthorized: Insufficient permissions to perform this action.');
-        this.errorMessage = 'Unauthorized: You do not have permission to perform this action.';
-        this.cdr.detectChanges(); // Manually trigger change detection
-        console.log('errorMessage: ' + this.errorMessage);
-        // Optionally, inform the user via UI by displaying this.errorMessage in your template
+        this.displayErrorAlert('Unauthorized: Insufficient permissions to perform this action.');
         return;
-      }
-  
-      const productData = this.productForm.value;
-      if (this.editingProductId) {
-        this.service.updateProduct(this.editingProductId, productData).then(() => {
-          console.log('Product updated successfully');
-          this.resetForm();
-        }).catch(error => {
-          console.error('Error updating product:', error);
-          this.errorMessage = 'Error updating product. Please try again.';
-        });
       } else {
-        this.service.addProduct(productData).then(() => {
-          console.log('Product added successfully');
-          this.resetForm();
-        }).catch(error => {
-          console.error('Error adding product:', error);
-          this.errorMessage = 'Error adding product. Please try again.';
-        });
+        if (this.productForm.valid) {
+          const productData = this.productForm.value;
+          if (this.editingProductId) {
+            this.service.updateProduct(this.editingProductId, productData).then(() => {
+              console.log('Product updated successfully');
+              this.resetForm();
+            }).catch(error => {
+              console.error('Error updating product:', error);
+              this.displayErrorAlert('Error updating product. Please try again.');
+            });
+          } else {
+            this.service.addProduct(productData).then(() => {
+              console.log('Product added successfully');
+              this.resetForm();
+            }).catch(error => {
+              console.error('Error adding product:', error);
+              this.displayErrorAlert('Error adding product. Please try again.');
+            });
+          }
+        } else {
+          console.error('Invalid Form');
+          this.displayErrorAlert('Error saving product. Please try again.');
+        }
       }
+
     }).catch(error => {
       console.error('Error checking permissions:', error);
-      this.errorMessage = 'Error checking permissions. Please refresh the page or try again later.';
+      this.displayErrorAlert('Error checking permissions. Please refresh the page or try again later.');
     });
   }
   
@@ -141,4 +163,38 @@ export class ProductComponent implements OnInit {
     this.editingProductId = null; // Clear the editing state
     this.refreshProducts(); // Refresh the list to show the updated data
   }
+
+  focusOnForm() {
+    console.log("focus on form")
+    console.log(this.productFormElem)
+    if (this.productFormElem) {
+      this.productFormElem.nativeElement.scrollIntoView({ behavior: 'smooth', block: 'center' });
+    }
+    
+  }
+
+  private showAlert(message: string) {
+    const alertElement = document.getElementById('alertMessage');
+    if (alertElement) {
+      alertElement.innerHTML = `
+        <div class="alert alert-danger alert-dismissible fade show" role="alert">
+          ${message}
+          <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+        </div>
+      `;
+      const closeButton = alertElement.querySelector('.btn-close');
+      if (closeButton) {
+        closeButton.addEventListener('click', () => {
+          alertElement.innerHTML = ''; // Close the alert when the close button is clicked
+        });
+      }
+    }
+  }
+
+  private displayErrorAlert(message: string) {
+    this.showAlert(message);
+    this.cdr.detectChanges(); 
+    console.log('errorMessage: ' + message);
+  }
+
 }
